TIP-221: move completeness generator interface
