+++ src/Pim/Component/Catalog/Completeness/CompletenessGeneratorInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Doctrine;
+namespace Pim\Component\Catalog\Completeness;
 
 use Pim\Component\Catalog\Model\ChannelInterface;
 use Pim\Component\Catalog\Model\FamilyInterface;
